requirements.txt
chainlit==2.9.3
langchain_core==1.2.4
langchain_openai==1.1.6
langchain_community
langchain_text_splitters
langchain_huggingface
chromadb
pypdf
sentence-transformers
python-dotenv==1.2.1
numpy
ddgs

